refactor: move stream manifest to own file
