+++ core/src/manifest/contract.rs
@@ -12,6 +12,7 @@ use crate::{
         AddressDetails, ContractEventMapping, FilterDetails, IndexingContractSetup,
     },
     indexer::parse_topic,
+    manifest::stream::StreamsConfig,
 };
 
 #[derive(Debug, Serialize, Deserialize, Clone)]
@@ -197,19 +198,6 @@ impl DependencyEventTree {
 
         dependencies
     }
-}
-
-#[derive(Debug, Serialize, Deserialize, Clone)]
-pub struct SNSStreamConfig {
-    pub prefix_id: Option<String>,
-    pub topic_arn: String,
-    pub networks: Vec<String>,
-    pub events: Vec<String>,
-}
-
-#[derive(Debug, Serialize, Deserialize, Clone)]
-pub struct StreamsConfig {
-    pub sns: Option<Vec<SNSStreamConfig>>,
 }
 
 #[derive(Debug, Serialize, Deserialize, Clone)]
