+++ core/src/manifest/stream.rs
@@ -1,0 +1,14 @@
+use serde::{Deserialize, Serialize};
+
+#[derive(Debug, Serialize, Deserialize, Clone)]
+pub struct SNSStreamConfig {
+    pub prefix_id: Option<String>,
+    pub topic_arn: String,
+    pub networks: Vec<String>,
+    pub events: Vec<String>,
+}
+
+#[derive(Debug, Serialize, Deserialize, Clone)]
+pub struct StreamsConfig {
+    pub sns: Option<Vec<SNSStreamConfig>>,
+}
